release:all worked 02.04.2023

+++ alarmClockFeature/src/main/java/com/yes/alarmclockfeature/domain/usecase/SetAndPlayTracksToPlayerPlaylistUseCase.kt
@@ -17,7 +17,7 @@ class SetAndPlayTracksToPlayerPlaylistUseCase (
         return params?.let {
            playerRepository.setTracks(
                params.tracks,
-               settingsRepository.subscribeCurrentTrackIndex().first()
+               settingsRepository.subscribeCurrentTrackIndex().first().toInt()
            )
             playerRepository.play()
             DomainResult.Success(Unit)
